# anyharness/sdk/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnyHarness Runtime API",
    "description": "Local sidecar runtime for managing coding agent installations, credentials, and ACP sessions.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces/{workspace_id}/mobility/destroy-source": {
      "post": {
        "tags": [
          "mobility"
        ],
        "operationId": "destroy_workspace_mobility_source",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestroyWorkspaceMobilitySourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Destroyed the old workspace source materialization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestroyWorkspaceMobilitySourceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/mobility/export": {
      "post": {
        "tags": [
          "mobility"
        ],
        "operationId": "export_workspace_mobility_archive",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportWorkspaceMobilityArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workspace mobility archive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMobilityArchive"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/mobility/install": {
      "post": {
        "tags": [
          "mobility"
        ],
        "operationId": "install_workspace_mobility_archive",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallWorkspaceMobilityArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Archive installed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallWorkspaceMobilityArchiveResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/mobility/preflight": {
      "post": {
        "tags": [
          "mobility"
        ],
        "operationId": "preflight_workspace_mobility",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace mobility preflight",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMobilityPreflightResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/mobility/runtime-state": {
      "put": {
        "tags": [
          "mobility"
        ],
        "operationId": "update_workspace_mobility_runtime_state",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceMobilityRuntimeStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workspace mobility runtime state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMobilityRuntimeState"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DestroyWorkspaceMobilitySourceRequest": {
        "type": "object"
      },
      "DestroyWorkspaceMobilitySourceResponse": {
        "type": "object",
        "required": [
          "workspaceId",
          "sourceDestroyed"
        ],
        "properties": {
          "closedTerminalIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deletedSessionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceDestroyed": {
            "type": "boolean"
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
...
      "ExportWorkspaceMobilityArchiveRequest": {
        "type": "object",
        "properties": {
          "excludePaths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "InstallWorkspaceMobilityArchiveRequest": {
        "type": "object",
        "required": [
          "archive"
        ],
        "properties": {
          "archive": {
            "$ref": "#/components/schemas/WorkspaceMobilityArchive"
          }
        }
      },
      "InstallWorkspaceMobilityArchiveResponse": {
        "type": "object",
        "required": [
          "workspaceId",
          "sourceWorkspacePath",
          "baseCommitSha",
          "importedSessionIds",
          "appliedFileCount",
          "deletedFileCount",
          "importedAgentArtifactCount"
        ],
        "properties": {
          "appliedFileCount": {
            "type": "integer",
            "minimum": 0
          },
          "baseCommitSha": {
            "type": "string"
          },
          "deletedFileCount": {
            "type": "integer",
            "minimum": 0
          },
          "importedAgentArtifactCount": {
            "type": "integer",
            "minimum": 0
          },
          "importedSessionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceWorkspacePath": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "UpdateWorkspaceMobilityRuntimeStateRequest": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "handoffOpId": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "$ref": "#/components/schemas/WorkspaceMobilityRuntimeMode"
          }
        }
      },
...
      "WorkspaceMobilityArchive": {
        "type": "object",
        "required": [
          "sourceWorkspacePath",
          "repoRootPath",
          "baseCommitSha",
          "files"
        ],
        "properties": {
          "baseCommitSha": {
            "type": "string"
          },
          "branchName": {
            "type": [
              "string",
              "null"
            ]
          },
          "deletedPaths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMobilityFileEntry"
            }
          },
          "repoRootPath": {
            "type": "string"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMobilitySessionBundle"
            }
          },
          "sourceWorkspacePath": {
            "type": "string"
          }
        }
      },
...
      "WorkspaceMobilityPreflightResponse": {
        "type": "object",
        "required": [
          "workspaceId",
          "canMove",
          "runtimeState"
        ],
        "properties": {
          "archiveEstimatedBytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "baseCommitSha": {
            "type": [
              "string",
              "null"
            ]
          },
          "blockers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMobilityBlocker"
            }
          },
          "branchName": {
            "type": [
              "string",
              "null"
            ]
          },
          "canMove": {
            "type": "boolean"
          },
          "runtimeState": {
            "$ref": "#/components/schemas/WorkspaceMobilityRuntimeState"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMobilitySessionCandidate"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
      "WorkspaceMobilityRuntimeMode": {
        "type": "string",
        "enum": [
          "normal",
          "frozen_for_handoff",
          "remote_owned",
          "repair_blocked"
        ]
      },
      "WorkspaceMobilityRuntimeState": {
        "type": "object",
        "required": [
          "workspaceId",
          "mode",
          "updatedAt"
        ],
        "properties": {
          "handoffOpId": {
            "type": [
              "string",
              "null"
            ]
          },
          "mode": {
            "$ref": "#/components/schemas/WorkspaceMobilityRuntimeMode"
          },
          "updatedAt": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}